backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

redis==5.2.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

pydantic==2.10.3
pydantic-settings==2.6.1

python-json-logger==2.0.7

prometheus-client==0.21.0
email-validator>=2.0.0
dnspython>=2.0.0

python-multipart==0.0.9